net: dt-bindings: Introduce the Qualcomm IPQESS Ethernet switch
Add the DT binding for the IPQESS Ethernet Switch. It inherits from a basic
ethernet switch binding and adds two regmaps, a phandle and reset line for
the PSGMII, and a phandle to the MDIO bus.

Add the DT binding for the IPQESS Ethernet Controller. This is a simple
controller, only requiring the phy-mode, interrupts, clocks, and possibly a
MAC address setting.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>