dts: qcom: ipq4019: Add description for the IPQ4019 ESS EDMA and switch
The Qualcomm IPQ4019 includes a modified version of the QCA8K Ethernet
switch. Add support for this device.

The switch's CPU port is connected to the SoC through the internal IPQESS
Ethernet controller. Add support for this internal interface, which only
supports a specific internal interface mode for connection to the switch.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>