net: dsa: qca: Make QCA8K hardware library available globally
The Qualcomm QCA8K Ethernet switch is already supported in the kernel, as
an external DSA switch. However, the Qualcomm IPQ4019 SoC contains an
internal modified QCA8K switch that does not fit into the DSA model, since
it uses an out-of-band tagging protocol.

Move the qca8k.h header file out of the QCA8K DSA driver so that the
IPQ4019 QCA8K switchdev driver can include it.

Refactor qca8k-common into a separate module so that the IPQESS driver can
be built as a module.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>